registration-app/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openpyxl>=3.1.0
httpx>=0.26.0
python-multipart>=0.0.7
PyYAML>=6.0
kubernetes>=28.0.0
